Refactor wagmi connection utilities and fix import paths
