+++ apps/iframe/src/components/ConnectModal.tsx
@@ -1,11 +1,12 @@
 import { createUUID } from "@happy.tech/common"
-import { type ConnectionProvider, Msgs, type MsgsFromApp, WalletDisplayAction } from "@happy.tech/wallet-common"
+import type { ConnectionProvider, MsgsFromApp } from "@happy.tech/wallet-common"
+import { Msgs, WalletDisplayAction } from "@happy.tech/wallet-common"
 import { useMutation } from "@tanstack/react-query"
 import { cx } from "class-variance-authority"
 import { useCallback, useEffect, useState } from "react"
-import { FirebaseErrorCode, isFirebaseError } from "#src/connections/firebase/errors.ts"
-import { iframeID } from "#src/requests/utils.ts"
-import { patchTimeoutOff, signalClosed } from "#src/utils/walletState.ts"
+import { FirebaseErrorCode, isFirebaseError } from "#src/connections/firebase/errors"
+import { iframeID } from "#src/requests/utils"
+import { patchTimeoutOff, signalClosed } from "#src/utils/walletState"
 import { useConnectionProviders } from "../connections/initialize"
 import { appMessageBus } from "../services/eventBus"
 import { DotLinearMotionBlurLoader } from "./loaders/DotLinearMotionBlurLoader"
