+++ apps/iframe/src/connections/GoogleConnector.ts
@@ -1,5 +1,4 @@
 import { type HappyUser, WalletType } from "@happy.tech/wallet-common"
-import { connect as connectWagmi, disconnect as disconnectWagmi } from "@wagmi/core"
 import { type AuthProvider, GoogleAuthProvider } from "firebase/auth"
 import type { EIP1193Provider } from "viem"
 import { setUserWithProvider } from "#src/actions/setUserWithProvider.ts"
@@ -7,8 +6,7 @@ import { StorageKey, storage } from "#src/services/storage.ts"
 import { getChains } from "#src/state/chains.ts"
 import { grantPermissions } from "#src/state/permissions.ts"
 import { getAppURL } from "#src/utils/appURL.ts"
-import { config } from "#src/wagmi/config.ts"
-import { happyConnector } from "#src/wagmi/connector.ts"
+import { connectWagmi, disconnectWagmi } from "#src/wagmi/utils"
 import { FirebaseConnector } from "./firebase"
 import { googleLogo } from "./firebase/logos"
 
@@ -64,7 +62,7 @@ export class GoogleConnector extends FirebaseConnector {
         if (storage.get(StorageKey.HappyUser)?.type !== WalletType.Social) return
         try {
             // if wagmi wasn't previously successfully connected, this throws
-            await disconnectWagmi(config)
+            await disconnectWagmi()
         } catch {}
         setUserWithProvider(undefined, undefined)
     }
@@ -72,14 +70,14 @@ export class GoogleConnector extends FirebaseConnector {
     async onReconnect(user: HappyUser, provider: EIP1193Provider) {
         await this.addSupportedChains(provider)
         setUserWithProvider(user, provider)
-        await connectWagmi(config, { connector: happyConnector })
+        await connectWagmi()
     }
 
     async onConnect(user: HappyUser, provider: EIP1193Provider) {
         await this.addSupportedChains(provider)
         setUserWithProvider(user, provider)
         grantPermissions(getAppURL(), "eth_accounts")
-        await connectWagmi(config, { connector: happyConnector })
+        await connectWagmi()
     }
 
     private async addSupportedChains(provider: EIP1193Provider) {
