+++ apps/iframe/src/connections/InjectedProviderProxy.ts
@@ -9,7 +9,6 @@ import {
     type ProviderEventPayload,
 } from "@happy.tech/wallet-common"
 import { SafeEventEmitter } from "@happy.tech/wallet-common"
-import { connect as connectWagmi, disconnect as disconnectWagmi } from "@wagmi/core"
 import type { EIP1193Provider } from "viem"
 import { setUserWithProvider } from "#src/actions/setUserWithProvider.ts"
 import { iframeID } from "#src/requests/utils.ts"
@@ -19,7 +18,6 @@ import { grantPermissions } from "#src/state/permissions.ts"
 import { getUser } from "#src/state/user.ts"
 import { getAppURL, isStandaloneIframe } from "#src/utils/appURL.ts"
 import { createHappyUserFromWallet } from "#src/utils/createHappyUserFromWallet.ts"
-import { happyConnector } from "#src/wagmi/connector.ts"
 import { iframeProvider } from "#src/wagmi/provider.ts"
 
 /**
@@ -115,12 +113,12 @@ export class InjectedProviderProxy extends SafeEventEmitter {
             this.emit("accountsChanged", accounts)
 
             // We need to dynamically import here to avoid any circular dependencies with iframeProvider
-            const { config } = await import("#src/wagmi/config.ts")
+            const { connectWagmi, disconnectWagmi } = await import("#src/wagmi/utils")
             const user = getUser()
             if (!accounts.length) {
                 try {
                     // if wagmi wasn't previously successfully connected, this throws
-                    await disconnectWagmi(config)
+                    await disconnectWagmi()
                 } catch {}
                 // Logout from the wallet when the user disconnects the injected wallet from the standalone wallet.
                 setUserWithProvider(undefined, undefined)
@@ -129,7 +127,7 @@ export class InjectedProviderProxy extends SafeEventEmitter {
                 const _user = await createHappyUserFromWallet(user.provider, address)
                 setUserWithProvider(_user, InjectedProviderProxy.getInstance() as EIP1193Provider)
                 grantPermissions(getAppURL(), "eth_accounts")
-                await connectWagmi(config, { connector: happyConnector })
+                await connectWagmi()
             }
         })
     }
@@ -145,14 +143,14 @@ export class InjectedProviderProxy extends SafeEventEmitter {
         switch (req.payload.event) {
             case "accountsChanged": {
                 // We need to dynamically import here to avoid any circular dependencies with iframeProvider
-                const { config } = await import("#src/wagmi/config.ts")
+                const { connectWagmi, disconnectWagmi } = await import("#src/wagmi/utils")
                 const user = getUser()
                 if (Array.isArray(req.payload.params) && !req.payload.params.length) {
                     // on disconnect, logout. alternatively could revoke permissions, but not much
                     // point in that for injected wallets
                     try {
                         // if wagmi wasn't previously successfully connected, this throws
-                        await disconnectWagmi(config)
+                        await disconnectWagmi()
                     } catch {}
                     setUserWithProvider(undefined, undefined)
                 } else if (Array.isArray(req.payload.params) && user) {
@@ -160,7 +158,7 @@ export class InjectedProviderProxy extends SafeEventEmitter {
                     const _user = await createHappyUserFromWallet(user.provider, address)
                     setUserWithProvider(_user, InjectedProviderProxy.getInstance() as EIP1193Provider)
                     grantPermissions(getAppURL(), "eth_accounts")
-                    await connectWagmi(config, { connector: happyConnector })
+                    await connectWagmi()
                 }
             }
         }
